Architectury 1.21.1/dev

+++ src/main/java/org/patryk3211/powergrid/circuits/components/CapacitorComponent.java
@@ -52,7 +52,7 @@ public class CapacitorComponent extends OrientableComponent {
     public boolean tick(@NotNull PlacedComponent placed) {
         if(!placed.wires.isEmpty()) {
             // Make charge persistent
-            placed.set(CHARGE, placed.wires.get(0).potentialDifference());
+            placed.set(CHARGE, (float) placed.wires.get(0).potentialDifference());
         }
         return true;
     }
