+++ src/main/java/org/patryk3211/powergrid/electricity/battery/BatteryBlockEntity.java
@@ -66,7 +66,7 @@ public class BatteryBlockEntity extends ElectricBlockEntity {
      * @return Positive power draws energy, negative power recharges
      */
     public float calculatePower() {
-        return -sourceCoupling.getCurrent() * sourceCoupling.getVoltage();
+        return (float) (-sourceCoupling.getCurrent() * sourceCoupling.getVoltage());
     }
 
     @Override
