+++ src/main/java/org/patryk3211/powergrid/electricity/creative/CreativeSourceBlockEntity.java
@@ -128,9 +128,9 @@ public class CreativeSourceBlockEntity extends ElectricBlockEntity implements IH
 
     public float getValue() {
         if(voltageSource) {
-            return voltageSourceNode.getVoltage();
+            return (float) voltageSourceNode.getVoltage();
         } else {
-            return currentSourceNode.getCurrent();
+            return (float) currentSourceNode.getCurrent();
         }
     }
