+++ src/main/java/org/patryk3211/powergrid/electricity/crt/CRTBlockEntity.java
@@ -106,9 +106,9 @@ public class CRTBlockEntity extends ElectricBlockEntity {
         if(level.isClientSide) {
             // 100 mA is needed for max brightness.
             // 50 mA is the minimum current.
-            brightness[head] = Mth.clamp((I - 0.05f) / 0.05f, 0, 1.5f);
-            xPoints[head] = Mth.clamp(xDeflect.current() / 0.5f, -1, 1);
-            yPoints[head] = Mth.clamp(yDeflect.current() / 0.5f, -1, 1);
+            brightness[head] = (float) Mth.clamp((I - 0.05) / 0.05, 0, 1.5);
+            xPoints[head] = (float) Mth.clamp(xDeflect.current() / 0.5, -1, 1);
+            yPoints[head] = (float) Mth.clamp(yDeflect.current() / 0.5, -1, 1);
             head = (head + 1) % brightness.length;
         }
     }
