+++ src/main/java/org/patryk3211/powergrid/electricity/gauge/CurrentGaugeBlockEntity.java
@@ -95,7 +95,7 @@ public class CurrentGaugeBlockEntity extends GaugeBlockEntity {
 
     @Override
     public float getValue() {
-        return wire.current();
+        return (float) wire.current();
     }
 
     @Override
