+++ src/main/java/org/patryk3211/powergrid/electricity/sim/node/CurrentSourceNode.java
@@ -24,7 +24,7 @@ import java.util.List;
  * Warning! Current source nodes cannot be directly connected to transformer couplings.
  */
 public class CurrentSourceNode extends ElectricNode implements IStaticResidual {
-    protected float current;
+    protected double current;
 
     public CurrentSourceNode() {
 
@@ -45,17 +45,17 @@ public class CurrentSourceNode extends ElectricNode implements IStaticResidual {
     }
 
     @Override
-    public float getVoltage() {
-        return (float) getStateValue();
+    public double getVoltage() {
+        return getStateValue();
     }
 
     @Override
-    public void setCurrent(float current) {
+    public void setCurrent(double current) {
         this.current = current;
     }
 
     @Override
-    public float getCurrent() {
+    public double getCurrent() {
         return current;
     }
