+++ src/main/java/org/patryk3211/powergrid/electricity/sim/node/FloatingNode.java
@@ -20,7 +20,7 @@ public class FloatingNode extends ElectricNode {
     }
 
     @Override
-    public float getVoltage() {
-        return (float) getStateValue();
+    public double getVoltage() {
+        return getStateValue();
     }
 }
