+++ src/main/java/org/patryk3211/powergrid/electricity/sim/special/GeneratorCoupling.java
@@ -44,7 +44,7 @@ public class GeneratorCoupling extends VoltageSourceCoupling {
 
     public void tick(float newField) {
         if(isConverged())
-            rotor.applyTickForce(field * getCurrent());
+            rotor.applyTickForce((float) (field * getCurrent()));
 
         setField(newField);
         setVoltage(field * rotor.getAngularVelocityRadians());
