+++ src/main/java/org/patryk3211/powergrid/electricity/wire/WireEntity.java
@@ -40,7 +40,7 @@ public abstract class WireEntity extends BaseWireEntity {
 
     @Override
     public float current() {
-        return wire == null || !wire.isConverged() ? 0 : wire.current();
+        return wire == null || !wire.isConverged() ? 0 : (float) wire.current();
     }
 
     @Override
