+++ src/main/java/org/patryk3211/powergrid/kinetics/motor/ConstantSpeedMotorBlockEntity.java
@@ -141,7 +141,7 @@ public class ConstantSpeedMotorBlockEntity extends GeneratingKineticBlockEntity 
 
         if(!level.isClientSide || isVirtual()) {
             applyPower(coil);
-            avgSpeed += calculateSpeed(coil.power(), torque()) * Math.signum(coil.current());
+            avgSpeed += (float) (calculateSpeed(coil.power(), torque()) * Math.signum(coil.current()));
         }
         super.tick();
     }
