+++ src/main/java/org/patryk3211/powergrid/kinetics/plotter/PlotterBlockEntity.java
@@ -154,7 +154,7 @@ public class PlotterBlockEntity extends ElectricKineticBlockEntity {
         super.tick();
         if(!isSpeedRequirementFulfilled())
             return;
-        headTarget = Mth.clamp(wire.potentialDifference() / maxValue, -1, 1);
+        headTarget = Mth.clamp((float) (wire.potentialDifference() / maxValue), -1, 1);
         prevHeadPosition = headPosition;
         headPosition += (headTarget - headPosition) * 0.9f;
